Update deconvoluted Tool and Dockstore Tooli
* updated swagger client, fix tests
