+++ dockstore-integration-testing/src/test/java/io/dockstore/client/cli/SystemClientIT.java
@@ -44,6 +44,7 @@ import io.swagger.client.ApiException;
 import io.swagger.client.api.ContainersApi;
 import io.swagger.client.api.GAGHApi;
 import io.swagger.client.api.UsersApi;
+import io.swagger.client.model.DockstoreTool;
 import io.swagger.client.model.Group;
 import io.swagger.client.model.RegisterRequest;
 import io.swagger.client.model.SourceFile;
@@ -124,11 +125,11 @@ public class SystemClientIT {
         UsersApi usersApi = new UsersApi(client);
         User user = usersApi.getUser();
 
-        List<Tool> tools = usersApi.userContainers(user.getId());
+        List<DockstoreTool> tools = usersApi.userContainers(user.getId());
         assertTrue(tools.size() == 2);
 
         ContainersApi containersApi = new ContainersApi(client);
-        List<Tool> containerList = containersApi.allContainers();
+        List<DockstoreTool> containerList = containersApi.allContainers();
         assertTrue(containerList.size() > 1);
     }
 
@@ -136,7 +137,7 @@ public class SystemClientIT {
     public void testFailedContainerRegistration() throws ApiException, IOException, TimeoutException {
         ApiClient client = getWebClient();
         ContainersApi containersApi = new ContainersApi(client);
-        List<Tool> containers = containersApi.allRegisteredContainers();
+        List<DockstoreTool> containers = containersApi.allRegisteredContainers();
 
         assertTrue(containers.size() == 1);
 
@@ -146,7 +147,7 @@ public class SystemClientIT {
 
         assertTrue(containers.size() == 5);
 
-        Tool container = containersApi.getContainerByToolPath("quay.io/test_org/test2");
+        DockstoreTool container = containersApi.getContainerByToolPath("quay.io/test_org/test2");
         assertFalse(container.getIsRegistered());
 
         long containerId = container.getId();
@@ -162,20 +163,20 @@ public class SystemClientIT {
         ApiClient client = getAdminWebClient();
         ContainersApi containersApi = new ContainersApi(client);
 
-        Tool c = getContainer();
+        DockstoreTool c = getContainer();
 
         containersApi.registerManual(c);
     }
 
-    private Tool getContainer() {
-        Tool c = new Tool();
-        c.setMode(Tool.ModeEnum.MANUAL_IMAGE_PATH);
+    private DockstoreTool getContainer() {
+        DockstoreTool c = new DockstoreTool();
+        c.setMode(DockstoreTool.ModeEnum.MANUAL_IMAGE_PATH);
         c.setName("seqware_full");
         c.setName("seqware");
         c.setGitUrl("https://github.com/denis-yuen/test1");
         c.setDefaultDockerfilePath("/Dockerfile");
         c.setDefaultCwlPath("/Dockstore.cwl");
-        c.setRegistry(Tool.RegistryEnum.DOCKER_HUB);
+        c.setRegistry(DockstoreTool.RegistryEnum.DOCKER_HUB);
         c.setIsRegistered(true);
         c.setIsPublic(true);
         c.setValidTrigger(true);
@@ -205,9 +206,9 @@ public class SystemClientIT {
         ApiClient client = getAdminWebClient();
         ContainersApi containersApi = new ContainersApi(client);
 
-        Tool c = getContainer();
+        DockstoreTool c = getContainer();
 
-        final Tool container = containersApi.registerManual(c);
+        final DockstoreTool container = containersApi.registerManual(c);
         containersApi.registerManual(container);
     }
 
@@ -227,7 +228,7 @@ public class SystemClientIT {
         GAGHApi toolApi = new GAGHApi(client);
         ContainersApi containersApi = new ContainersApi(client);
         // register one more to give us something to look at
-        Tool c = getContainer();
+        DockstoreTool c = getContainer();
         containersApi.registerManual(c);
 
         List<Tool> tools = toolApi.toolsGet(null, null, null, null, null, null, null);
@@ -263,7 +264,7 @@ public class SystemClientIT {
         GAGHApi toolApi = new GAGHApi(client);
         ContainersApi containersApi = new ContainersApi(client);
         // register one more to give us something to look at
-        Tool c = getContainer();
+        DockstoreTool c = getContainer();
         containersApi.registerManual(c);
 
         final ToolDockerfile toolDockerfile = toolApi.toolsRegistryIdVersionVersionIdDockerfileGet("registry.hub.docker.com/seqware/seqware/test5","master");
@@ -276,7 +277,7 @@ public class SystemClientIT {
     public void testContainerRegistration() throws ApiException, IOException, TimeoutException {
         ApiClient client = getWebClient();
         ContainersApi containersApi = new ContainersApi(client);
-        List<Tool> containers = containersApi.allRegisteredContainers();
+        List<DockstoreTool> containers = containersApi.allRegisteredContainers();
 
         assertTrue(containers.size() == 1);
 
@@ -286,7 +287,7 @@ public class SystemClientIT {
 
         assertTrue(containers.size() == 5);
 
-        Tool container = containersApi.getContainerByToolPath("quay.io/test_org/test5");
+        DockstoreTool container = containersApi.getContainerByToolPath("quay.io/test_org/test5");
         assertFalse(container.getIsRegistered());
 
         long containerId = container.getId();
@@ -311,7 +312,7 @@ public class SystemClientIT {
         ApiClient client = getWebClient();
         ContainersApi containersApi = new ContainersApi(client);
 
-        List<Tool> containers = containersApi.search("test6");
+        List<DockstoreTool> containers = containersApi.search("test6");
         assertTrue(containers.size() == 1);
         assertTrue(containers.get(0).getPath().equals("quay.io/test_org/test6"));
 
@@ -325,7 +326,7 @@ public class SystemClientIT {
 
         ContainersApi containersApi = new ContainersApi(client);
         // register one more to give us something to look at
-        Tool c = getContainer();
+        DockstoreTool c = getContainer();
         c.getTags().get(0).setHidden(true);
         c = containersApi.registerManual(c);
 
@@ -333,7 +334,7 @@ public class SystemClientIT {
 
         ApiClient muggleClient = getWebClient();
         ContainersApi muggleContainersApi = new ContainersApi(muggleClient);
-        final Tool registeredContainer = muggleContainersApi.getRegisteredContainer(c.getId());
+        final DockstoreTool registeredContainer = muggleContainersApi.getRegisteredContainer(c.getId());
         assertTrue("should see no tags as a regular user, saw " + registeredContainer.getTags().size(), registeredContainer.getTags().size() == 0);
     }
